split api class connection by file
