+++ src/services/api/base.ts
@@ -1,0 +1,75 @@
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import axios from 'axios';
+import { STORAGE_USER_AUTH_TOKEN } from 'helpers/constants';
+import { DevSettings } from 'react-native';
+import config from '../../../config';
+
+axios.defaults.baseURL = config.baseApiUrl;
+
+async function getRequest<T>(
+    endpoint: string,
+    useAuthToken = false
+): Promise<T | undefined> {
+    let response: T | undefined;
+    try {
+        let result;
+        if (useAuthToken) {
+            const token = await AsyncStorage.getItem(STORAGE_USER_AUTH_TOKEN);
+            const requestOptions = {
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                },
+            };
+            result = await axios.get(endpoint, requestOptions);
+        } else {
+            result = await axios.get(endpoint);
+        }
+        if (result.status >= 400) {
+            return undefined;
+        }
+        if (result.data === '') {
+            // TODO: this condition should not exist
+            response = undefined;
+        } else {
+            response = result.data as T;
+        }
+    } catch (error) {
+        // Api.system.uploadError('', 'get_request', error);
+    }
+    return response;
+}
+
+async function postRequest<T>(
+    endpoint: string,
+    requestBody: any,
+    options?: any
+): Promise<T | undefined> {
+    let response: T | undefined;
+    try {
+        // handle success
+        const token = await AsyncStorage.getItem(STORAGE_USER_AUTH_TOKEN);
+        const requestOptions = {
+            headers: {
+                Authorization: `Bearer ${token}`,
+                'Content-Type': 'application/json',
+                Accept: 'application/json',
+            },
+            ...options,
+        };
+        const result = await axios.post(endpoint, requestBody, requestOptions);
+        if (result.status === 401) {
+            await AsyncStorage.clear();
+            DevSettings.reload();
+            return undefined;
+        }
+        if (result.status >= 400) {
+            return undefined;
+        }
+        response = result.data as T;
+    } catch (error) {
+        // Api.system.uploadError('', 'post_request', error);
+    }
+    return response;
+}
+
+export { getRequest, postRequest };
